v1.1 build 2 changes
Realized I left in the ColorOptionsVC but it didn’t serve any purpose
anymore because the new tip system made those options obsolete.

So to “remove” the VC from the app:
1. I disabled the Options Bar Button in ColorMenuVC
2. Removed any picture or title the button may have had
3. Commented out the tip regarding the Options button in the ColorMenuVC

This way, I can cleanly remove work I’ve already done and reintegrate
it later if need be.

Will archive the app like this and submit it to the app store for v1.1
release